src/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.4.0
async-timeout==4.0.2
asyncpg==0.25.0
asynctest==0.13.0
atlastk==0.13
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
ffmpeg-python==0.2.0
frozenlist==1.2.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httptools==0.3.0
httpx==0.18.2
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
iso8601==0.1.16
MarkupSafe==2.0.1
multidict==5.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.6
pypika-tortoise==0.1.1
pytest==6.2.5
pytest-asyncio==0.16.0
pytz==2021.3
requests==2.28.0
rfc3986==1.5.0
sanic==21.9.3
sanic-routing==0.7.2
sanic-session==0.8.0
sanic-testing==0.8.2
sniffio==1.2.0
tomli==2.0.1
tomlkit==0.11.0
soupsieve==2.3.1
SQLAlchemy==1.4.39
toml==0.10.2
tortoise-orm==0.17.8
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.7
uvloop==0.16.0
websockets==10.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
